Comp2001 web project/requirements.txt
alembic==1.14.0
bcml==3.10.8
blinker==1.8.2
botw-utils==0.2.3
cefpython3==66.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.7
clr-loader==0.2.5
colorama==0.4.6
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.4
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.1.4  # Corrected version of Mako
MarkupSafe==2.1.5
oead==1.2.6.post1
packaging==21.3
proxy-tools==0.1.0
pycparser==2.21
pyodbc==5.2.0
pyparsing==3.0.9
python-dotenv==1.0.1
pythonnet==3.0.1
pywebview==3.7.2
PyYAML==6.0.2
requests==2.27.1
rstb==1.2.2
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==1.26.15
Werkzeug==3.0.6
xxhash==3.2.0
zipp==3.20.2
flask-restx==1.3.0
PyJWT==2.9.0
flask-swagger==0.2.14
flask-swagger-ui==4.11.1